--- accessibility-api/src/main/java/co/vividhata/accessibility_api/model/PageCheck.java
@@ -1,5 +1,0 @@
-package co.vividhata.accessibility_api.model;
-
-import java.time.Instant;
-
-public record PageCheck(int pageCheckId, int webPageId, Instant timeChecked, String htmlContent) { }
