+++ accessibility-api/src/main/java/co/vividhata/accessibility_api/model/Scan.java
@@ -1,0 +1,5 @@
+package co.vividhata.accessibility_api.model;
+
+import java.time.Instant;
+
+public record Scan(int pageCheckId, int webPageId, Instant timeScanned, String htmlContent) { }
