--- accessibility-api/src/main/java/co/vividhata/accessibility_api/read_page/IPageCheckRepository.java
@@ -1,9 +1,0 @@
-package co.vividhata.accessibility_api.read_page;
-
-import java.time.Instant;
-
-public interface IPageCheckRepository {
-
-    int create(int webPageId, Instant timeChecked, String htmlContent);
-
-}
