--- accessibility-api/src/main/java/co/vividhata/accessibility_api/read_page/IReadPageService.java
@@ -1,7 +1,0 @@
-package co.vividhata.accessibility_api.read_page;
-
-public interface IReadPageService {
-
-    String readPageFrom(String url, int accountId);
-
-}
