+++ accessibility-api/src/main/java/co/vividhata/accessibility_api/scan/IScanRepository.java
@@ -1,0 +1,9 @@
+package co.vividhata.accessibility_api.scan;
+
+import java.time.Instant;
+
+public interface IScanRepository {
+
+    int create(int webPageId, Instant timeScanned, String htmlContent);
+
+}
