+++ accessibility-api/src/main/java/co/vividhata/accessibility_api/scan/IScanService.java
@@ -1,0 +1,7 @@
+package co.vividhata.accessibility_api.scan;
+
+public interface IScanService {
+
+    String readPageFrom(String url, int accountId);
+
+}
