+++ accessibility-api/src/main/java/co/vividhata/accessibility_api/scan/IWebPageRepository.java
@@ -1,4 +1,4 @@
-package co.vividhata.accessibility_api.read_page;
+package co.vividhata.accessibility_api.scan;
 
 import co.vividhata.accessibility_api.model.WebPage;
