+++ accessibility-api/src/main/java/co/vividhata/accessibility_api/scan/PostgreSqlScanRepository.java
@@ -1,6 +1,5 @@
-package co.vividhata.accessibility_api.read_page;
+package co.vividhata.accessibility_api.scan;
 
-import co.vividhata.accessibility_api.model.WebPage;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.support.GeneratedKeyHolder;
@@ -8,24 +7,27 @@ import org.springframework.jdbc.support.KeyHolder;
 import org.springframework.stereotype.Service;
 
 import java.sql.PreparedStatement;
+import java.sql.Timestamp;
+import java.time.Instant;
 
 @Service
-public class PostgreSqlWebPageRepository implements IWebPageRepository {
+public class PostgreSqlScanRepository implements IScanRepository {
 
     @Autowired
-    JdbcTemplate jdbcTemplate;
+    private JdbcTemplate jdbcTemplate;
 
     @Override
-    public int create(int accountId, String url) {
-        String sql = "INSERT INTO ac.web_page(account_id, url) VALUES (?, ?);";
+    public int create(int webPageId, Instant timeScanned, String htmlContent) {
+        String sql = "INSERT INTO ac.scan(web_page_id, time_scanned, html_content) VALUES (?, ?, ?);";
 
         KeyHolder keyHolder = new GeneratedKeyHolder();
 
         jdbcTemplate.update(con -> {
 
             PreparedStatement ps = con.prepareStatement(sql, new String[]{"id"});
-            ps.setInt(1, accountId);
-            ps.setString(2, url);
+            ps.setInt(1, webPageId);
+            ps.setTimestamp(2, Timestamp.from(timeScanned));
+            ps.setString(3, htmlContent);
             return ps;
 
         }, keyHolder);
@@ -36,11 +38,4 @@ public class PostgreSqlWebPageRepository implements IWebPageRepository {
         }
         return id.intValue();
     }
-
-    public WebPage get(int accountId, String url) {
-        String sql = "SELECT * FROM ac.web_page WHERE account_id = ? AND url = ?;";
-
-        return jdbcTemplate.query(sql, WebPage::fromResultSet, accountId, url);
-    }
-
 }
